Update registration.php
I took care of as many issues with the page as I saw fit. I have JS determining the size of a drop down menu and in conjunction with a sql query that checks to see if the username has already been used when they leave the text box. I didn't bother implementing encryption because I didn't know how we wanted to do that. Also I didn't actually check to see if the queries worked because I didn't feel like creating the whole database as of tonight. Let me know if it works. 